requirements.txt
langchain-google-genai>=0.0.3
browser-use>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0